+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 1.3.0 (Unreleased)
 
-- **Added:** `examples` [GH-13](https://github.com/terraform-alicloud-modules/terraform-alicloud-slb/pull/13)
+IMPROVEMENTS:
+
+- improve readme and variables [GH-14](https://github.com/terraform-alicloud-modules/terraform-alicloud-slb/pull/14)
+- improve and add examples [GH-13](https://github.com/terraform-alicloud-modules/terraform-alicloud-slb/pull/13)